+++ src/twitux-network.h
@@ -50,7 +50,7 @@ void twitux_network_logout 		(void);
 
 /* Post a new tweet */
 void twitux_network_post_status		(const gchar *text, 
-					gint reply_id);
+					gint64 reply_id);
 
 /* Post a direct message to a follower */
 void twitux_network_send_message	(const gchar *friend,
